+++ Libraries/LibWebView/CMakeLists.txt
@@ -13,7 +13,6 @@ set(SOURCES
     HelperProcess.cpp
     Menu.cpp
     Mutation.cpp
-    Plugins/FontPlugin.cpp
     Plugins/ImageCodecPlugin.cpp
     Process.cpp
     ProcessHandle.cpp
